Fixed many references to UTC-like behaviour. Renamed Region to DateRegion Split up into separate files Consistent file naming Consolidated many initialisation code into one.
TODO: mocking string tests
TODO: all comments in the headers
TODO: Documentation
TODO: Quick & Nimble
TODO: calendar and time zone extensions are not elegant enough
