+++ SwiftDate/DateInRegionComparisons.swift
@@ -1,0 +1,113 @@
+//
+//  DateInRegionComparisons.swift
+//  Pods
+//
+//  Created by Jeroen Houtzager on 26/10/15.
+//
+//
+
+import Foundation
+
+// MARK: - Comparators
+
+public extension DateInRegion {
+
+    /// Returns an NSComparisonResult value that indicates the ordering of two given dates based on their components down to a given unit granularity.
+    ///
+    /// - Parameters:
+    ///     - date: date to compare.
+    ///     - toUnitGranularity: The smallest unit that must, along with all larger units, be equal for the given dates to be considered the same.
+    ///         For possible values, see “[Calendar Units](xcdoc://?url=developer.apple.com/library/prerelease/ios/documentation/Cocoa/Reference/Foundation/Classes/NSCalendar_Class/index.html#//apple_ref/c/tdef/NSCalendarUnit)”
+    ///
+    /// - Returns: NSOrderedSame if the dates are the same down to the given granularity, otherwise NSOrderedAscending or NSOrderedDescending.
+    ///
+    /// - seealso: [compareDate:toDate:toUnitGranularity:](xcdoc://?url=developer.apple.com/library/prerelease/ios/documentation/Cocoa/Reference/Foundation/Classes/NSCalendar_Class/index.html#//apple_ref/occ/instm/NSCalendar/compareDate:toDate:toUnitGranularity:)
+    ///
+    public func compareDate(date: DateInRegion, toUnitGranularity unit: NSCalendarUnit) -> NSComparisonResult {
+        return calendar.compareDate(self.absoluteTime, toDate: date.absoluteTime, toUnitGranularity: unit)
+    }
+
+    /// Returns whether the given date is in today.
+    ///
+    /// - Returns: a boolean indicating whether the receiver is in today
+    ///
+    /// - note: This value is interpreted in the context of the calendar of the receiver
+    ///
+    /// - seealso: [isDateInToday:](xcdoc://?url=developer.apple.com/library/prerelease/ios/documentation/Cocoa/Reference/Foundation/Classes/NSCalendar_Class/index.html#//apple_ref/occ/instm/NSCalendar/isDateInToday:)
+    ///
+    public func isInToday(  ) -> Bool {
+        return calendar.isDateInToday(absoluteTime)
+    }
+
+    /// Returns whether the given date is in yesterday.
+    ///
+    /// - Returns: a boolean indicating whether the receiver is in yesterday
+    ///
+    /// - note: This value is interpreted in the context of the calendar of the receiver
+    ///
+    /// - seealso: [isDateInYesterday:](xcdoc://?url=developer.apple.com/library/prerelease/ios/documentation/Cocoa/Reference/Foundation/Classes/NSCalendar_Class/index.html#//apple_ref/occ/instm/NSCalendar/isDateInYesterday:)
+    ///
+    public func isInYesterday() -> Bool {
+        return calendar.isDateInYesterday(absoluteTime)
+    }
+
+    /// Returns whether the given date is in tomorrow.
+    ///
+    /// - Returns: a boolean indicating whether the receiver is in tomorrow
+    ///
+    /// - note: This value is interpreted in the context of the calendar of the receiver
+    ///
+    /// - seealso: [isDateInTomorrow:](xcdoc://?url=developer.apple.com/library/prerelease/ios/documentation/Cocoa/Reference/Foundation/Classes/NSCalendar_Class/index.html#//apple_ref/occ/instm/NSCalendar/isDateInTomorrow:)
+    ///
+    public func isInTomorrow() -> Bool {
+        return calendar.isDateInTomorrow(absoluteTime)
+    }
+
+    /// Returns whether the given date is in the weekend.
+    ///
+    /// - Returns: a boolean indicating whether the receiver is in the weekend
+    ///
+    /// - note: This value is interpreted in the context of the calendar of the receiver
+    ///
+    /// - seealso: [isDateInWeekend:](xcdoc://?url=developer.apple.com/library/prerelease/ios/documentation/Cocoa/Reference/Foundation/Classes/NSCalendar_Class/index.html#//apple_ref/occ/instm/NSCalendar/isDateInWeekend:)
+    ///
+    public func isInWeekend() -> Bool {
+        return calendar.isDateInWeekend(absoluteTime)
+    }
+
+}
+
+// MARK: - Comparable delegate
+
+
+/// Instances of conforming types can be compared using relational operators, which define a strict total order.
+///
+/// A type conforming to Comparable need only supply the < and == operators; default implementations of <=, >, >=, and != are supplied by the standard library:
+///
+extension DateInRegion: Comparable {}
+
+/// Returns whether the given date is later than the receiver.
+/// Just the dates are compared. Calendars, time zones are irrelevant.
+///
+/// - Parameters:
+///     - date: a date to compare against
+///
+/// - Returns: a boolean indicating whether the receiver is earlier than the given date
+///
+public func <(ldate: DateInRegion, rdate: DateInRegion) -> Bool {
+    return ldate.absoluteTime.compare(rdate.absoluteTime) == .OrderedAscending
+}
+
+/// Returns whether the given date is earlier than the receiver.
+/// Just the dates are compared. Calendars, time zones are irrelevant.
+///
+/// - Parameters:
+///     - date: a date to compare against
+///
+/// - Returns: a boolean indicating whether the receiver is later than the given date
+///
+public func >(ldate: DateInRegion, rdate: DateInRegion) -> Bool {
+    return ldate.absoluteTime.compare(rdate.absoluteTime) == .OrderedDescending
+}
+
+
