+++ SwiftDate/DateInRegionHashable.swift
@@ -1,0 +1,13 @@
+//
+//  DateInRegionHashable.swift
+//  Pods
+//
+//  Created by Jeroen Houtzager on 04/11/15.
+//
+//
+
+extension DateInRegion: Hashable {
+    public var hashValue: Int {
+        return absoluteTime.hashValue ^ region.hashValue
+    }
+}
